+++ test/unit/transaction_test.js
@@ -1,4 +1,6 @@
 import randomBytes from 'randombytes';
+import { doesNotReject } from 'assert';
+import { afterEach } from 'mocha';
 
 describe('Transaction', function() {
   it('constructs Transaction object from a TransactionEnvelope', function(done) {
@@ -455,6 +457,143 @@ describe('Transaction', function() {
     expect(typeof transaction).to.be.equal('object');
     expect(typeof transaction.toXDR).to.be.equal('function');
     expect(transaction.toXDR()).to.be.equal(xdrString);
+  });
+
+  describe('.buildFeeBumpTransaction', function() {
+    it('builds a fee bump transaction', function(done) {
+      let networkPassphrase = 'Standalone Network ; February 2017';
+      let innerSource = StellarBase.Keypair.master(networkPassphrase);
+      let innerAccount = new StellarBase.Account(innerSource.publicKey(), '7');
+      let destination =
+        'GDQERENWDDSQZS7R7WKHZI3BSOYMV3FSWR7TFUYFTKQ447PIX6NREOJM';
+      let amount = '2000.0000000';
+
+      let innerTx = new StellarBase.TransactionBuilder(innerAccount, {
+        fee: 100,
+        networkPassphrase: networkPassphrase,
+        timebounds: {
+          minTime: 0,
+          maxTime: 0
+        },
+        v1: true
+      })
+        .addOperation(
+          StellarBase.Operation.payment({
+            destination,
+            asset: StellarBase.Asset.native(),
+            amount
+          })
+        )
+        .addMemo(StellarBase.Memo.text('Happy birthday!'))
+        .build();
+
+      innerTx.sign(innerSource);
+      let feeSource = StellarBase.Keypair.fromSecret(
+        'SB7ZMPZB3YMMK5CUWENXVLZWBK4KYX4YU5JBXQNZSK2DP2Q7V3LVTO5V'
+      );
+      let baseFee = '100';
+      let transaction = StellarBase.Transaction.buildFeeBumpTransaction(
+        feeSource,
+        baseFee,
+        innerTx.toEnvelope().value(),
+        networkPassphrase
+      );
+      transaction.sign(feeSource);
+
+      const expectedXDR =
+        'AAAABQAAAADgSJG2GOUMy/H9lHyjYZOwyuyytH8y0wWaoc596L+bEgAAAAAAAADIAAAAAgAAAABzdv3ojkzWHMD7KUoXhrPx0GH18vHKV0ZfqpMiEblG1gAAAGQAAAAAAAAACAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAQAAAA9IYXBweSBiaXJ0aGRheSEAAAAAAQAAAAAAAAABAAAAAOBIkbYY5QzL8f2UfKNhk7DK7LK0fzLTBZqhzn3ov5sSAAAAAAAAAASoF8gAAAAAAAAAAAERuUbWAAAAQK933Dnt1pxXlsf1B5CYn81PLxeYsx+MiV9EGbMdUfEcdDWUySyIkdzJefjpR5ejdXVp/KXosGmNUQ+DrIBlzg0AAAAAAAAAAei/mxIAAABAijIIQpL6KlFefiL4FP8UWQktWEz4wFgGNSaXe7mZdVMuiREntehi1b7MRqZ1h+W+Y0y+Z2HtMunsilT2yS5mAA==';
+
+      expect(transaction.isFeeBump()).to.equal(true);
+      expect(
+        transaction
+          .toEnvelope()
+          .toXDR()
+          .toString('base64')
+      ).to.be.equal(expectedXDR);
+
+      expect(transaction.source).to.be.equal(innerAccount.accountId());
+      expect(transaction.feeSource).to.be.equal(feeSource.publicKey());
+
+      // shows new fee
+      expect(transaction.fee).to.be.equal('200');
+
+      // can read inner fee
+      expect(transaction.innerFee).to.be.equal('100');
+
+      // show innerTx operations and memo
+      let operation = transaction.operations[0];
+      expect(transaction.memo.type).to.be.equal(StellarBase.MemoText);
+      expect(transaction.memo.value.toString('ascii')).to.be.equal(
+        'Happy birthday!'
+      );
+      expect(operation.type).to.be.equal('payment');
+      expect(operation.destination).to.be.equal(destination);
+      expect(operation.amount).to.be.equal(amount);
+
+      let expectedTxEnvelope = StellarBase.xdr.TransactionEnvelope.fromXDR(
+        expectedXDR,
+        'base64'
+      ).value();
+
+      expect(transaction.source).to.equal(
+        StellarBase.StrKey.encodeEd25519PublicKey(
+          expectedTxEnvelope
+            .tx()
+            .innerTx()
+            .value()
+            .tx()
+            .sourceAccount()
+            .ed25519()
+        )
+      );
+      expect(transaction.feeSource).to.equal(
+        StellarBase.StrKey.encodeEd25519PublicKey(
+          expectedTxEnvelope
+            .tx()
+            .feeSource()
+            .ed25519()
+        )
+      );
+
+      expect(transaction.innerFee).to.equal(
+        expectedTxEnvelope
+          .tx()
+          .innerTx()
+          .value()
+          .tx()
+          .fee()
+          .toString()
+      );
+      expect(transaction.fee).to.equal(
+        expectedTxEnvelope
+          .tx()
+          .fee()
+          .toString()
+      );
+
+      expect(transaction.innerSignatures.length).to.equal(1);
+      expect(
+        transaction.innerSignatures[0].toXDR().toString('base64')
+      ).to.equal(
+        expectedTxEnvelope
+          .tx()
+          .innerTx()
+          .value()
+          .signatures()[0]
+          .toXDR()
+          .toString('base64')
+      );
+
+      expect(transaction.signatures.length).to.equal(1);
+      expect(transaction.signatures[0].toXDR().toString('base64')).to.equal(
+        expectedTxEnvelope
+          .signatures()[0]
+          .toXDR()
+          .toString('base64')
+      );
+
+      done();
+    });
   });
 });
